fix: name the file in the ambient status label for file tools
describeToolCall read toolInput.file_path, but the file tools emit
toolInput.fileName (emitFileToolCall in agent/tools/text-editor.ts), so
every file_write/file_edit/file_batch_edit/file_read run fell back to the
generic "Editing files" / "Reading files" instead of naming the file. The
visualizer's toolDisplay.ts already reads fileName; only the host-side
vocabulary was wrong.

Read fileName first and keep file_path as a fallback — tools that pass
their raw Zod input straight through, such as migration_source_read and
the MCP bridge, do carry the schema's file_path.

Moved the mapping into its own leaf module so it can be tested:
importing AgentStatusManager pulls in VisualizerWebview and transitively
an ESM-only LS dependency that Jest cannot parse. Mirrors the split
already used for AgentStreamView/toolDisplay.ts on the webview side.

The new test fails on 6 of 13 cases against the old field read.
